adding baseball prospectus scraper + data intake added as part of historical refresh, mongodb updated with WAR data from 1951-present day

+++ projects/SABR/refresh.py
@@ -12,7 +12,6 @@ import os
 import argparse
 
 CUR_SEASON = "2019"
-HIST_SEASON_RANGE = np.arange(1990, 2020)
 
 def conn():
 	return pymongo.MongoClient("mongodb+srv://admin:pdometer@mongo-uwij2.mongodb.net/test?retryWrites=true")
@@ -22,7 +21,8 @@ def refresh_table(db_name, table_name, df):
 	client = conn()
 	db = client[db_name]
 	table = db[table_name]
-	table.drop()
+	if table.count() != 0:
+		table.drop()
 	table.insert(json.loads(data_json))
 
 def insert_to_table(db_name, table_name, df):
@@ -40,10 +40,14 @@ def load_teams():
 	table.drop()
 	table.insert(df)
 
-def load_team_historical():
-	for season in HIST_SEASON_RANGE:
-		df = hist.get_finished_df(season)
-		refresh_table('MLB_TEAM_HISTORICAL', 'fg_dashboard_'+str(season), df)
+def load_team_historical(source, season_range):
+	for season in season_range:
+		if source == 'fg':
+			df = hist.get_finished_df(season, 'fg')
+			refresh_table('MLB_TEAM_HISTORICAL', 'fg_dashboard_'+str(season), df)
+		elif source == 'bp':
+			df = hist.get_finished_df(season, 'bp')
+			refresh_table('MLB_TEAM_HISTORICAL', 'bp_team_warp_'+str(season), df)
 
 def load_bref_team_sp():
 	url = "https://www.baseball-reference.com/leagues/MLB/{}-starter-pitching.shtml".format(CUR_SEASON)
@@ -69,22 +73,47 @@ def load_opener_candidates():
 	insert_to_table('SABR', 'opener_candidates', rsp_df)
 	insert_to_table('SABR', 'opener_candidates', lsp_df)
 
-def main():
-	parser = argparse.ArgumentParser(description='SABR Daily Data Refresh')
-	parser.add_argument('--hist', action='store_true', help="Choose whether to run the historical data refresh or not")
-	args = parser.parse_args()
-
+def main(args):
 	if args.hist:
-		print("### RUNNING HISTORICAL LOAD ###")
-		load_teams()
-		load_team_historical()
-		print("### HISTORICAL LOAD COMPLETES ###")
-	print("### RUNNING DAILY LOAD ###")
-	load_bref_team_sp()
-	load_batter_fp()
-	load_pitcher_fp()
-	load_opener_candidates()
-	print("### DAILY LOAD COMPLETES ###")
+		if args.start_year and args.end_year:
+			season_range = np.arange(int(args.start_year), int(args.end_year))
+		elif args.start_year:
+			season_range = np.arange(int(args.start_year), int(CUR_SEASON)+1)
+		elif args.end_year:
+			if args.bp:
+				season_range = np.arange(1951, int(args.end_year))
+			else:
+				season_range = np.arange(1903, int(args.end_year))
+		else:
+			season_range = np.arange(1903, int(CUR_SEASON)+1)
+		
+		if args.bp:
+			load_teams()
+			load_team_historical('bp', season_range)
+		elif args.fg:
+			load_teams()
+			load_team_historical('fg', season_range)
+		else:
+			print("### RUNNING HISTORICAL LOAD ###")
+			load_teams()
+			load_team_historical('fg', season_range)
+			load_team_historical('bp', season_range)
+			print("### HISTORICAL LOAD COMPLETES ###")
+	if args.daily:
+		print("### RUNNING DAILY LOAD ###")
+		load_bref_team_sp()
+		load_batter_fp()
+		load_pitcher_fp()
+		load_opener_candidates()
+		print("### DAILY LOAD COMPLETES ###")
 
 if __name__ == '__main__':
-	main()
+	parser = argparse.ArgumentParser(description='SABR Daily Data Refresh')
+	parser.add_argument('--daily', action='store_true', help="Choose whether to run the daily data refresh or not")
+	parser.add_argument('--hist', action='store_true', help="Choose whether to run the historical data refresh or not")
+	parser.add_argument('--bp', action='store_true', help="Only run bp section of the historical refresh")
+	parser.add_argument('--fg', action='store_true', help="Only run fg section of the historical refresh")
+	parser.add_argument('--start_year', help="Specify a start year for historical refresh")
+	parser.add_argument('--end_year', help="Specify an end year for historical refresh")
+	args = parser.parse_args()
+	main(args)
